moved filterTrailList to its own file
