+++ app/api/data/data.js
@@ -1,12 +1,4 @@
 import { supabase } from "@/app/api/data/initSupabase";
-import {
-  ANY_AREA,
-  ANY_DIFFICULTY,
-  ANY_LENGTH,
-  SHORT,
-  MEDIUM,
-  LONG,
-} from "@/app/lib/constants";
 import dayjs from "dayjs";
 import utc from "dayjs/plugin/utc";
 import timezone from "dayjs/plugin/timezone";
@@ -45,29 +37,6 @@ export async function addTrail(trailInfo) {
     throw new Error("Failed to add new trail.");
   }
   return data;
-}
-
-export async function filterTrailList(area_name, difficulty_rating, length) {
-  let newList = await fetchAllTrails();
-  if (area_name !== ANY_AREA) {
-    newList = newList.filter((trail) => trail.area_name == area_name);
-  }
-  if (difficulty_rating !== ANY_DIFFICULTY) {
-    newList = newList.filter(
-      (trail) => trail.difficulty_rating == difficulty_rating
-    );
-  }
-  if (length !== ANY_LENGTH) {
-    newList = newList.filter(
-      (trail) =>
-        (length == SHORT && Number(trail.length) < 3) ||
-        (length == LONG && Number(trail.length > 6)) ||
-        (length == MEDIUM &&
-          Number(trail.length) >= 3 &&
-          Number(trail.length) <= 6)
-    );
-  }
-  return newList;
 }
 
 //User functions
